+++ activesupport/lib/active_support/digest.rb
@@ -1,19 +1,29 @@
 # frozen_string_literal: true
 
+require "openssl"
+
 module ActiveSupport
   class Digest #:nodoc:
     class <<self
       def hash_digest_class
-        @hash_digest_class ||= ::Digest::MD5
+        @hash_digest_class ||= "MD5"
       end
 
       def hash_digest_class=(klass)
-        raise ArgumentError, "#{klass} is expected to implement hexdigest class method" unless klass.respond_to?(:hexdigest)
+        unless klass.is_a?(String)
+          klass = klass.to_s.split("::").last.to_s
+          ActiveSupport::Deprecation.warn(<<~EOM)
+            Passing a Digest Class to `hash_digest_class=` is deprecated and will be removed in future
+            versions of Rails. Please pass a string instead (SHA1, SHA256, MD5, etc).
+          EOM
+        end
+
+        raise ArgumentError, "#{klass} is unsupported by OpenSSL::Digest" unless OpenSSL::Digest.const_defined?(klass)
         @hash_digest_class = klass
       end
 
       def hexdigest(arg)
-        hash_digest_class.hexdigest(arg)[0...32]
+        OpenSSL::Digest.hexdigest(hash_digest_class, arg)[0...32]
       end
     end
   end